+++ src/components/stakePanel/StakePanel.jsx
@@ -156,11 +156,7 @@ const StakePanel = ({ state, openModal }) => {
         
 
   }
-  const unstake = async () => {
-    await pool.exit().catch((e) => {
-      console.log("insufficientBalance", e);
-    });
-  };
+  
 
   const exitInactivePools = () => {
     state.manager.exitInactive();
@@ -212,13 +208,6 @@ const StakePanel = ({ state, openModal }) => {
           <button className="button" disabled={!state.provider} onClick={stake}>
             stake
           </button>
-          {/* <button
-            className="button"
-            disabled={!state.provider}
-            onClick={unstake}
-          >
-            unstake
-          </button> */}
           
           {inactivePools.length > 0 && (
             <button
